Picker View changes

+++ Source/Helpers/YPPhotoSaver.swift
@@ -46,20 +46,20 @@ public class YPPhotoSaver {
         return nil
     }
     
-    func clearAllFile() {
+   class func clearAllFile() {
         let fileManager = FileManager.default
 
         let paths = NSSearchPathForDirectoriesInDomains(.documentDirectory, .userDomainMask, true).first!
-
         print("Directory: \(paths)")
-
         do {
             let fileName = try fileManager.contentsOfDirectory(atPath: paths)
 
             for file in fileName {
                 // For each file in the directory, create full path and delete the file
-                let filePath = URL(fileURLWithPath: paths).appendingPathComponent(file).absoluteURL
-                try fileManager.removeItem(at: filePath)
+                if file == YPConfig.albumName{
+                    let filePath = URL(fileURLWithPath: paths).appendingPathComponent(file).absoluteURL
+                    try fileManager.removeItem(at: filePath)
+                }
             }
         } catch {
             print(error.localizedDescription)
